donations_project/requirements.txt
Django==4.2
djangorestframework==3.14.0
python-decouple==3.5
psycopg2-binary==2.9.6
django-cors-headers==3.14.0
```